[Fix] avoid using offset [Add] geometry for components
The nexus-constructor does not handle the offset attribute of a
transformation correctly. By splitting the reduced affine transformation
into a translation and rotation eniius can write compatible JSON NeXus
structures until nexus-constructor is fixed.

Nexus-constructor can draw the geometry of components, but requires
conversion to OFF to do so.
The Elliptic_guide_gravity component is added as a method of
McStasComp2NX with the *guide surfaces* specifed in as an OFF geometry.

The DiskChopper component is moved from NX2COMP_MAP to a McStasComp2NX
method in order to calculate parameters so that nexus-constructor can
display a representation of the disk. At present the position of the
disk is wrong in the nexus-constructor view since NeXus and McStas
disagree on the reference point of the disk (the disk centre vs the slit
opening centre, respectively).