[Refactor] `rotmat` from Euler angles, `get_euler_angles`
All comments regarding the rotation matrix convention used in McStas now
reside in the `rotmat` method. The component ROTATED angles are again
input as-is, with the transformation matrix now representing the
coordinate transformation directly (all sin(*) terms are negated, but
even numbered occurences cancel out).

The `get_euler_angles` method (which may be unused) is updated to
follow the same convention as the calculated `rotmat`. This may not be
necessary, but is done for consistency.