Bugfixes for nexus-constructor (#13)
* Bugfixes for nexus-constructor (#13)

Squashed commit of the following:

commit 13d818c7b16ccb2a94922c861a19d0f5dfdced43
Author: Duc Le <duc.le@stfc.ac.uk>
Date:   Mon Jul 31 18:48:23 2023 +0100

    Fix sense of rotation on input from McStas

commit 8a388a18da2a3f6308dd99d9803f681fea6a7218
Author: Duc Le <duc.le@stfc.ac.uk>
Date:   Mon Jul 31 15:25:52 2023 +0100

    Fix error in matrix multiplication order in reduce_affice_transforms()

commit 2c60d35350e34b1d70286b27e5dcabea3193ff8b
Author: Duc Le <duc.le@stfc.ac.uk>
Date:   Mon Jul 31 14:31:19 2023 +0100

    Bugfixes identified in testing with nexus-constructor

    Fix error in computing Euler angles from rotation matrix
      (thanks @g5t)
    Add check to AffineRotate constructor to check that
      input matrix is valid
    Fix error in computing the inverse transformation (@g5t)

* [Fix] avoid using offset [Add] geometry for components

The nexus-constructor does not handle the offset attribute of a
transformation correctly. By splitting the reduced affine transformation
into a translation and rotation eniius can write compatible JSON NeXus
structures until nexus-constructor is fixed.

Nexus-constructor can draw the geometry of components, but requires
conversion to OFF to do so.
The Elliptic_guide_gravity component is added as a method of
McStasComp2NX with the *guide surfaces* specifed in as an OFF geometry.

The DiskChopper component is moved from NX2COMP_MAP to a McStasComp2NX
method in order to calculate parameters so that nexus-constructor can
display a representation of the disk. At present the position of the
disk is wrong in the nexus-constructor view since NeXus and McStas
disagree on the reference point of the disk (the disk centre vs the slit
opening centre, respectively).

* [Fix] dangling reference to renamed parameter

* [Refactor] `rotmat` from Euler angles, `get_euler_angles`

All comments regarding the rotation matrix convention used in McStas now
reside in the `rotmat` method. The component ROTATED angles are again
input as-is, with the transformation matrix now representing the
coordinate transformation directly (all sin(*) terms are negated, but
even numbered occurences cancel out).

The `get_euler_angles` method (which may be unused) is updated to
follow the same convention as the calculated `rotmat`. This may not be
necessary, but is done for consistency.

---------

Co-authored-by: Gregory Tucker <gregory.tucker@ess.eu>